+++ Excel_PRIME.Tests/RowCellTestsXlsb.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading.Tasks;
@@ -27,8 +28,10 @@ public class RowCellTestsXlsb
             using ISheetAsync? worksheet = await workbook.GetSheetAsync(sheetName).ConfigureAwait(false);
             await foreach (IRowAsync? row in worksheet!.GetRowDataAsync().ConfigureAwait(false))
             {
-                await foreach (ICell? cell in row.GetAllCellsAsync().ConfigureAwait(false))
+                IReadOnlyList<ICell?>? rowCells = await row.GetAllCellsAsync().ConfigureAwait(true);
+                foreach (ICell? cell in rowCells)
                 {
+                    // Because this returns upto the dimension of the sheet width
                     if (cell == null)
                     {
                         // Because this returns upto the dimension of the sheet width
@@ -65,8 +68,10 @@ public class RowCellTestsXlsb
         await foreach (IRowAsync? row in worksheet1.GetRowDataAsync().ConfigureAwait(false))
         {
             int c = 0;
-            await foreach (ICell? cell in row!.GetAllCellsAsync().ConfigureAwait(false))
+            IReadOnlyList<ICell?>? rowCells = await row.GetAllCellsAsync().ConfigureAwait(true);
+            foreach (ICell? cell in rowCells)
             {
+                // Because this returns upto the dimension of the sheet width
                 if (cell == null) // Because this returns upto the dimension of the sheet width
                 {
                     break;
@@ -107,8 +112,10 @@ public class RowCellTestsXlsb
         await foreach (IRowAsync? row in worksheet2.GetRowDataAsync().ConfigureAwait(false))
         {
             int c = 0;
-            await foreach (ICell? cell in row!.GetAllCellsAsync().ConfigureAwait(false))
+            IReadOnlyList<ICell?>? rowCells = await row.GetAllCellsAsync().ConfigureAwait(true);
+            foreach (ICell? cell in rowCells)
             {
+                // Because this returns upto the dimension of the sheet width
                 if (c > 0 && cell == null) // Because this returns upto the dimension of the sheet width
                 {
                     break;
@@ -153,8 +160,10 @@ public class RowCellTestsXlsb
             await foreach (IRowAsync? row in worksheet1.GetRowDataAsync().ConfigureAwait(false))
             {
                 int c = 0;
-                await foreach (ICell? cell in row!.GetAllCellsAsync().ConfigureAwait(false))
+                IReadOnlyList<ICell?>? rowCells = await row.GetAllCellsAsync().ConfigureAwait(true);
+                foreach (ICell? cell in rowCells)
                 {
+                    // Because this returns upto the dimension of the sheet width
                     if (cell == null) // Because this returns upto the dimension of the sheet width
                     {
                         break;
@@ -197,8 +206,10 @@ public class RowCellTestsXlsb
             await foreach (IRowAsync? row in worksheet2.GetRowDataAsync().ConfigureAwait(false))
             {
                 int c = 0;
-                await foreach (ICell? cell in row!.GetAllCellsAsync().ConfigureAwait(false))
+                IReadOnlyList<ICell?>? rowCells = await row.GetAllCellsAsync().ConfigureAwait(true);
+                foreach (ICell? cell in rowCells)
                 {
+                    // Because this returns upto the dimension of the sheet width
                     if (c > 0 && cell == null) // Because this returns upto the dimension of the sheet width
                     {
                         break;
@@ -244,15 +255,20 @@ public class RowCellTestsXlsb
         await foreach (IRowAsync? row in worksheet2.GetRowDataAsync().ConfigureAwait(false))
         {
             int c = 0;
-            await foreach (ICell? cell in row!.GetAllCellsAsync().ConfigureAwait(false))
+            IReadOnlyList<ICell?>? rowCells = await row.GetAllCellsAsync().ConfigureAwait(true);
+            if (rowCells != null)
             {
-                if (cell == null) // Because this returns upto the dimension of the sheet width
+                foreach (ICell? cell in rowCells)
                 {
-                    continue;
-                }
+                    // Because this returns upto the dimension of the sheet width
+                    if (cell == null) // Because this returns upto the dimension of the sheet width
+                    {
+                        continue;
+                    }
 
-                cell.CellValue.BoxedValue.Should().Be(workSheet2Content[r][c]);
-                c++;
+                    cell.CellValue.BoxedValue.Should().Be(workSheet2Content[r][c]);
+                    c++;
+                }
             }
 
             c.Should().Be(workSheet2Content[r].Length);