+++ Release_Notes.md
@@ -1,10 +1,14 @@
 ﻿# 2026-01-04 - V3 XLSB-Beta
+- ⛓️‍💥 **Breaking Change(s)
+    - Change `GetAllCells` to return `IReadOnlyList<ICell?>?`
 - Add `NonClosingStream` to  make it self documenting
 - Remove `BufferedStream` usage from xlsx
-- Add more XLSB readers
+- Add more XLSB Benchmark readers
+- Performance improvements 🚀
     - [2026-01-04](https://github.com/Smurf-IV/Excel_PRIME/blob/main/Performance.md#2026-01-04)
 - Remove some compile warnings
-
+- Make use of `[MethodImpl(MethodImplOptions.AggressiveOptimization)]`
+ 
 # 2026-01-02 - V3 XLSB-Beta
 - ⛓️‍💥 **Breaking Change(s)
     - Removal of the Conversion options `Number###`